\name{HI}
\alias{HI}
\docType{data}
\title{Data from Deutshbauer et al. (2005)}
\description{
  Mechanisms of Haploinsufficiency revealed by Genome-Wide Profiling in
  Yeast (Deutshbauer et al.,2005)
}
\usage{data(HI)}
\details{
  HI stands for haploinsufficient. The dataframe is composed of:
  \item{orf}{Yeast ORF, systematic name}
  \item{gene}{Yeast common gene name of the corresponding ORF}
  \item{go}{GO terms}
}
\source{
  \url{http://www.sciencemag.org/cgi/data/303/5659/808/DC1/1}
}
\references{
  Deutschbauer AM, Jaramillo DF, Proctor M, Kumm J, Hillenmeyer ME, Davis RW,
Nislow C, Giaever G. (2005) Mechanisms of haploinsufficiency revealed by genome-wide profiling in yeast.
 Genetics. 169(4):1915-25.
 }
\examples{
data(HI)
}
\keyword{datasets}
