\name{PCpheno-package}
\alias{PCpheno-package}
\alias{PCpheno}
\docType{package}
\title{
Linkage between Protein Complexes, Pathways, and Phenotypes
}
\description{
  Tools to integrate, annotate and search for associations between
  phenotypes, protein complexes, and pathways.
  }
\details{
\tabular{ll}{
Package: \tab PCpheno\cr
Type: \tab Package\cr
Version: \tab 1.3.1\cr
Date: \tab 2006-03-09\cr
License: \tab  The Artistic License, Version 2.0\cr
}

}
\author{N. LeMeur and R. Gentleman

Maintainer: N. LeMeur <nlemeur@fhcrc.org>

}
\references{
Giaever G, et al.(2002) Functional profiling of the Saccharomyces cerevisiae
genome. Nature. 418(6896):387-91. PMID: 12140549 \cr

Deutschbauer AM, Jaramillo DF, Proctor M, Kumm J, Hillenmeyer ME, Davis RW,
Nislow C, Giaever G. (2005) Mechanisms of haploinsufficiency revealed by genome-wide profiling in yeast.
Genetics. 169(4):1915-25.\cr

Byrne KP, Wolfe KH. (2005) The Yeast Gene Order Browser: combining curated homology and syntenic context
reveals gene fate in polyploid species. Genome Res. 15(10):1456-61. PMID: 16169922 

}
\keyword{package}
\seealso{ScISI,SLGI}
